edited pics and made a 1 page layout i like better
I edited pictures to all have transparent backgrounds. I was able to get the pages to be  on one document and don't need the extra pages. That is why i removed the other. I want to come up with a different contact page than originally planned on. Going to do some searching for ideas before I complete the final page.